refactor(tts): update TypeScript configuration and import paths
- Modified tsconfig.json to set the root directory to 'src' and adjusted include/exclude paths for better clarity.
- Updated import statements in runner modules to reflect the new directory structure, enhancing code organization and maintainability.
