+++ modules/stops/apps/tts/src/index.ts
@@ -1,7 +1,7 @@
 /* * */
 
-import { runnerCommon } from '@/src/runners/runner_common.js';
-import { runnerStops } from '@/src/runners/runner_stops.js';
+import { runnerCommon } from '@/runners/runner_common.js';
+import { runnerStops } from '@/runners/runner_stops.js';
 import { Logger } from '@tmlmobilidade/logger';
 import { Timer } from '@tmlmobilidade/timer';
 import { runOnInterval } from '@tmlmobilidade/utils';
