+++ modules/stops/apps/tts/src/runners/runner_common.ts
@@ -1,6 +1,6 @@
 /* * */
 
-import { generatePiperTtsAudio } from '@/src/services/piperTtsApi.js';
+import { generatePiperTtsAudio } from '@/services/piperTtsApi.js';
 import TIMETRACKER from '@helperkits/timer';
 import { files } from '@tmlmobilidade/interfaces';
 import { Logger } from '@tmlmobilidade/logger';
