+++ modules/stops/apps/tts/src/runners/runner_patterns.ts
@@ -1,6 +1,6 @@
 /* * */
 
-import { makePattern } from '@/src/utils/makeText.js';
+import { makePattern } from '@/utils/makeText.js';
 import TIMETRACKER from '@helperkits/timer';
 import { type HubLine, type HubPattern } from '@tmlmobilidade/go-types-public-info';
 import { Logger } from '@tmlmobilidade/logger';
