+++ modules/stops/apps/tts/tsconfig.json
@@ -3,10 +3,10 @@
 	"compilerOptions": {
 		"outDir": "dist",
 		"paths": {
-			"@/*": ["./*"]
+			"@/*": ["./src/*"]
 		},
-		"rootDir": "."
+		"rootDir": "src"
 	},
-	"include": ["."],
-	"exclude": ["../node_modules", "../dist"]
+	"include": ["src"],
+	"exclude": ["node_modules", "dist"]
 }
